completed another lab on coursera, adding it here
This lab is focused on working with libraries, especially pandas, requests, and BeautifulSoup. It also includes APIs, YahooFinance in this case. 

During the lab I had to extract stock data from Yahoo, and parse revenue table from a website. As a result, the graph was produced representin revenue over time. The graph is not made by me, it was precoded by the lab's authors